requirements.txt
numpy
opencv-python 
Pillow 
scikit-image 
scipy 
tensorboard 
torch==1.8.0
torchvision==0.9.0
pandas
tensorboard

cmake
gdown
ipykernel
ipywidgets
joblib
jsonargparse==3.19.4
jsonschema
jupyterlab
matplotlib
numba>=0.50
progress
scikit-learn>=0.22.2.post1
seaborn
umap-learn>=0.4.5
yapf